continue with operations

+++ 10_Projects/06_PySASSExt/py_sass_ext/src/Utils.hpp
@@ -20,7 +20,7 @@
 
 namespace SASS {
     using FArgString = std::string;
-    using FArgInt = int;
+    using FArgInt = int64_t;
     using FArgSASSBits = SASS_Bits;
     using FArgBool = bool;
     using FArgFloat = float;
